refactor: rename cloud function to cloud function gen2 (#23)
