+++ src/spaceone/inventory/conf/cloud_service_conf.py
@@ -25,7 +25,7 @@ CLOUD_SERVICE_GROUP_MAP = {
         'SubscriptionManager',
         'TopicManager'
     ],
-    'CloudFunctions': ['FunctionManager']
+    'CloudFunctions': ['FunctionGen2Manager', 'FunctionGen1Manager']
 }
 
 REGION_INFO = {
