+++ src/spaceone/inventory/connector/cloud_functions/function_gen2.py
@@ -2,11 +2,11 @@ import logging
 
 from spaceone.inventory.libs.connector import GoogleCloudConnector
 
-__all__ = ['FunctionConnector']
+__all__ = ['FunctionGen2Connector']
 _LOGGER = logging.getLogger(__name__)
 
 
-class FunctionConnector(GoogleCloudConnector):
+class FunctionGen2Connector(GoogleCloudConnector):
     google_client_service = 'cloudfunctions'
     version = 'v2'
