+++ src/spaceone/inventory/manager/__init__.py
@@ -16,5 +16,6 @@ from spaceone.inventory.manager.pub_sub.schema_manager import SchemaManager
 from spaceone.inventory.manager.pub_sub.snapshot_manager import SnapshotManager
 from spaceone.inventory.manager.pub_sub.subscription_manager import SubscriptionManager
 from spaceone.inventory.manager.pub_sub.topic_manager import TopicManager
-from spaceone.inventory.manager.cloud_functions.function_manager import FunctionManager
+from spaceone.inventory.manager.cloud_functions.function_gen2_manager import FunctionGen2Manager
+from spaceone.inventory.manager.cloud_functions.function_gen1_manager import FunctionGen1Manager
