+++ src/spaceone/inventory/model/cloud_functions/__init__.py
@@ -1,1 +1,2 @@
-from spaceone.inventory.model.cloud_functions.function import *
+from spaceone.inventory.model.cloud_functions.function_gen2 import *
+from spaceone.inventory.model.cloud_functions.function_gen1 import *
