--- src/spaceone/inventory/model/cloud_functions/function/inner_data/__init__.py
@@ -1,4 +1,0 @@
-from spaceone.inventory.model.cloud_functions.function.inner_data.build_config import *
-from spaceone.inventory.model.cloud_functions.function.inner_data.service_config import *
-from spaceone.inventory.model.cloud_functions.function.inner_data.event_trigger import *
-from spaceone.inventory.model.cloud_functions.function.inner_data.state_message import *
