+++ src/spaceone/inventory/model/cloud_functions/function_gen2/inner_data/__init__.py
@@ -1,0 +1,4 @@
+from spaceone.inventory.model.cloud_functions.function_gen2.inner_data.build_config import *
+from spaceone.inventory.model.cloud_functions.function_gen2.inner_data.service_config import *
+from spaceone.inventory.model.cloud_functions.function_gen2.inner_data.event_trigger import *
+from spaceone.inventory.model.cloud_functions.function_gen2.inner_data.state_message import *
