+++ src/spaceone/inventory/model/cloud_functions/function_gen2/inner_data/build_config.py
