+++ src/spaceone/inventory/model/cloud_functions/function_gen2/inner_data/event_trigger.py
