+++ src/spaceone/inventory/model/cloud_functions/function_gen2/widget/count_by_region.yml
