+++ src/spaceone/inventory/model/cloud_functions/function_gen2/widget/total_count.yml
